Comentar Implementado.
Pequeños cambios de diseño.

+++ src/java/BusinessLogic/Score/IScoreService.java
@@ -13,7 +13,7 @@ import Entities.User;
  * @author Usuario
  */
 public interface IScoreService {
-    public float GetPromObject(String objectType, int objectId);
+    public int GetPromObject(String objectType, int objectId);
     public void SaveScore(User objUser, int score, String objType, int objId);
     public void SaveScore(Score objScore);
     public Score GetScoreByObjeto(String objectType, int objectId, int userId);
