+++ src/java/DataAccess/Score/IScoreRepository.java
@@ -12,7 +12,7 @@ import Entities.Score;
  * @author Usuario
  */
 public interface IScoreRepository {
-    public float CalcularPromObjeto(String objectType, int objectId);
+    public int CalcularPromObjeto(String objectType, int objectId);
     public void SaveScore(Score objScore);
     public Score GetScoreByObjeto(String objectType, int objectId, int userId);
     public void UpdateScore(Score objScore);
